refactor: move get_tyndp_res_carriers to _helpers.py
