+++ scripts/add_brownfield.py
@@ -15,11 +15,12 @@ import xarray as xr
 from scripts._helpers import (
     configure_logging,
     get_snapshots,
+    get_tyndp_res_carriers,
     sanitize_custom_columns,
     set_scenario_config,
     update_config_from_wildcards,
 )
-from scripts.add_electricity import flatten, get_tyndp_res_carriers, sanitize_carriers
+from scripts.add_electricity import flatten, sanitize_carriers
 from scripts.add_existing_baseyear import add_build_year_to_new_assets
 
 logger = logging.getLogger(__name__)
